backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
pytest
httpx